feat: 타이머 페이지 sse 연결 유지
