+++ src/App.vue
@@ -35,7 +35,7 @@
 import { computed, ref, watch } from "vue";
 import { reactive, provide } from "vue";
 import { useRoute } from "vue-router";
-import { useStore } from "vuex";
+import { EventSourcePolyfill } from "event-source-polyfill";
 import store from "./store/store";
 import LoginModal from "./components/modal/LoginPermissionRequired.vue";
 import Sidebar from "./components/Sidebar.vue";
@@ -69,6 +69,7 @@ async function fetchGroups() {
     globalState.myGroups = [];
   }
 }
+
 // 로그인 상태 확인
 const isLoggedIn = computed(() => store.state.isLoggedIn);
 const isUserListVisible = ref(false);
@@ -98,6 +99,106 @@ const state = reactive({
 });
 // 상태 제공
 provide("state", state);
+
+
+/* SSE 연결 관리 */
+let eventSource = null; // SSE 전역 변수
+let reconnectAttempts = 0;
+const maxReconnectAttempts = 5;
+const groupId = ref(null);
+const users = reactive({ list: [] });
+const token = localStorage.getItem("access");
+
+const connectSSE = async () => {
+  if (eventSource) return; // 기존 SSE가 있으면 중복 연결 방지
+
+  eventSource = new EventSourcePolyfill(
+    `${import.meta.env.VITE_SERVER_HOST}/groups/${groupId.value}/subscribe`,
+    {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    }
+  );
+
+  eventSource.addEventListener("connect", (event) => {
+    reconnectAttempts = 0;
+    console.log("SSE connect:", event.data);
+  });
+
+  eventSource.addEventListener("heartbeat", (event) => {
+    console.log("heartbeat 수신:", event.data);
+  });
+
+  eventSource.onerror = (error) => {
+    console.error("SSE 연결 오류:", error);
+    eventSource.close();
+    eventSource = null;
+    store.commit("clearUsers");
+
+    if (reconnectAttempts < maxReconnectAttempts) {
+      const retryTime = Math.pow(2, reconnectAttempts) * 1000; // 지수 백오프 (1s, 2s, 4s, 8s, 16s)
+      console.log(`SSE 재연결 시도 ${reconnectAttempts + 1}회 후 ${retryTime / 1000}초 대기...`);
+
+      setTimeout(() => {
+        reconnectAttempts++;
+        connectSSE();
+      }, retryTime);
+    } else {
+      console.warn("SSE 최대 재연결 횟수를 초과했습니다.");
+    }
+  };
+
+  eventSource.addEventListener("memberData", (event) => {
+    try {
+      users.list = JSON.parse(event.data);
+      store.commit("setUsers", users);
+    } catch (error) {
+      console.error("데이터 파싱 오류:", error);
+    }
+  });
+
+  eventSource.addEventListener("statusUpdate", (event) => {
+    const data = JSON.parse(event.data);
+    store.commit("updateStatus", data);
+  });
+
+  return eventSource;
+};
+
+const closeSSE = () => {
+  if (eventSource && typeof eventSource.close === "function") {
+    console.log("SSE 연결 종료");
+    eventSource.close();
+    eventSource = null;
+    // Vuex 상태 초기화
+    store.commit("clearUsers");
+  }
+};
+
+// 라우트 변경 감지하여 SSE 관리
+watch(
+  () => route.fullPath,
+  (newPath) => {
+    const match = newPath.match(/^\/groups\/(\d+)/); // `/groups/:groupId` 패턴 검사
+    if (match) {
+      const newGroupId = match[1];
+
+      if (groupId.value !== newGroupId) {
+        console.log(`🔄 그룹 변경 감지: ${groupId.value} → ${newGroupId}`);
+        closeSSE();
+        groupId.value = newGroupId;
+        connectSSE();
+      }
+    } else {
+      console.log("🏠 그룹 페이지가 아님 - SSE 연결 해제");
+      closeSSE();
+      groupId.value = null;
+    }
+  },
+  { immediate: true }
+);
+
 </script>
 <script>
 import Header from "./components/Header.vue";
